add auto-refresh / split Vuex store into modules
